Extract main.js
