+++ package.json
@@ -2,12 +2,11 @@
   "name": "pr-summary",
   "version": "2.0.0",
   "description": "Summarize Pull Request between specified refs",
-  "main": "src/pr-summary.js",
+  "main": "src/main.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
     "lint": "eslint",
     "prepublish": "babel src -d lib --source-maps inline",
-    "js": "babel-node src/pr-summary.js"
+    "js": "babel-node src/main.js"
   },
   "repository": {
     "type": "git",
@@ -21,7 +20,7 @@
     "lib"
   ],
   "bin": {
-    "pr-summary" : "./lib/pr-summary.js"
+    "pr-summary" : "./lib/main.js"
   },
   "author": "Hiroki Kondo",
   "license": "MIT",
