+++ src/pr-summary.js
@@ -3,7 +3,6 @@
 
 import GitHubAPI from 'github';
 import 'babel-polyfill';
-import program from 'commander';
 
 const GITHUB_TOKEN = process.env.GITHUB_TOKEN;
 if (!GITHUB_TOKEN) {
@@ -66,7 +65,7 @@ const fetchPullRequests = (owner, repo, prNumbers) => {
   return Promise.all(getPRInfos);
 };
 
-const getPRInfo = (owner, repo, number) => {
+export const getPRInfo = (owner, repo, number) => {
   return new Promise((resolve) => {
     const prInfo = {};
     getCommitsFromPullRequest(owner, repo, number).
@@ -82,40 +81,4 @@ const getPRInfo = (owner, repo, number) => {
       });
   });
 };
-
-const prsummary = (owner, repo, number) => {
-  getPRInfo(owner, repo, number).
-    then((prInfo) => {
-      console.info('## Contributors\n');
-      const rendered = [];
-      prInfo.commits.map((commit) => {
-        const user = commit.author.login;
-        if (rendered.includes(user) === false && user.trim() !== '') {
-          console.info(user);
-          rendered.push(user);
-        }
-      });
-
-      console.info('');
-
-      console.info('## Pull Requests\n');
-      prInfo.prs.map((pr) => {
-        if (pr.pull_request) {
-          const message = `#${pr.number} ${pr.title} by ${pr.user.login}`;
-          console.info(message);
-        }
-      });
-    });
-};
-
-program.version('2.0.0').
-  usage('<owner> <repo> <number>').
-  action(prsummary);
-
-program.parse(process.argv);
-
-if(program.args.length < 3) {
-  program.help();
-}
-
